requirements.txt
pickle
opencv-python
ultralytics
supervision
numpy
Pillow
easyocr
gtts
playsound
TTS
clip_interrogator
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118